requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
apscheduler==3.10.4
httpx==0.28.1
feedparser==6.0.11
python-dotenv==1.0.1
jinja2==3.1.5
aiofiles==24.1.0
